Elevator Teleporting
Added elevator teleportation. Crew are now aware of elevator positions and will teleport accordingly. Need to fix strange bug on why they teleport twice.

+++ Assets/Scripts/Object_Script.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class Object_Script : MonoBehaviour
 {
 
-    public float currentLoc = 0;
+    public int currentLoc = 0;
     // Start is called before the first frame update
     void Start()
     {
@@ -20,6 +20,6 @@ public class Object_Script : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        
+
     }
 }
